fix: moving converter functions to aux file to not interfere with css properties
